Improve wheel input and GUI scroll redraw
Switch mouse wheel delta from integer to float across the app event path so high-resolution wheels and touchpad gestures keep fractional notches instead of being quantized. In GUI region layout, force a redraw after wheel-driven scroll updates so retained rendering does not appear to miss every other notch when no other input occurs. Sandbox logging/demo code was updated to show fractional deltas and provide enough content to exercise scrolling, and local tooling settings plus commented legacy vector typedef blocks in math.h were included in the changeset.

+++ source/engine/app/app.h
@@ -235,8 +235,8 @@ typedef struct app_mouse_btn_event_s /* 8 bytes */
 
 typedef struct app_mouse_wheel_event_s /* 8 bytes */
 {
-    i32 delta; /* signed; positive = toward user            */
-    i16 x, y;  /* client-area cursor position               */
+    f32 delta; /* signed notches (fractional); positive = toward user */
+    i16 x, y;  /* client-area cursor position                         */
 
 } app_mouse_wheel_event_t;
